backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework_simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv